+++ apps/web/src/components/signup/index.ts
@@ -1,8 +1,11 @@
-export { default as Agreement } from './agreement'
-export { default as Id } from './id'
-export { default as Name } from './name'
-export { default as Password } from './password'
-export { default as PhoneNumber } from './phonenumber'
-export { default as Birthdate } from './birthdate'
-export { default as Sex } from './sex'
-export { default as SignupDone } from './signup-done'
+export { default as Agreement } from './step/agreement'
+export { default as Id } from './step/id'
+export { default as Name } from './step/name'
+export { default as Password } from './step/password'
+export { default as PhoneNumber } from './step/phonenumber'
+export { default as Birthdate } from './step/birthdate'
+export { default as Sex } from './step/sex'
+export { default as SignupDone } from './step/signup-done'
+export * from './signup-title'
+export * from './type'
+export * from './with-form-context'
