+++ apps/web/src/components/signup/step/birthdate.tsx
@@ -1,9 +1,11 @@
 import { isValidDate } from '@/utils/is-valid-date'
 import { Button, TextField } from '@yaksok/ui'
 import { useWatch } from 'react-hook-form'
-import SignupTitle from './signup-title'
-import { WithFormContext } from './type'
-import { withFormContext } from './with-form-context'
+import {
+  SignupTitle,
+  withFormContext,
+  type WithFormContext,
+} from '@components/signup'
 
 function BirthDate({ methods, title, onNext }: WithFormContext) {
   const birthDateValue = useWatch({
