+++ apps/web/src/components/signup/step/name.tsx
@@ -1,9 +1,11 @@
 import { nameRegex, SignupRequest } from '@/validation/zod'
 import { Button, TextField } from '@yaksok/ui'
 import { useWatch } from 'react-hook-form'
-import SignupTitle from './signup-title'
-import { WithFormContext } from './type'
-import { withFormContext } from './with-form-context'
+import {
+  SignupTitle,
+  withFormContext,
+  type WithFormContext,
+} from '@components/signup'
 
 function Name({ methods, title, onNext }: WithFormContext) {
   const { control, register, handleSubmit } = methods
