+++ apps/web/src/components/signup/step/password.tsx
@@ -1,9 +1,11 @@
 import { passwordRegex } from '@/validation/zod'
 import { Button, TextField } from '@yaksok/ui'
 import { useWatch } from 'react-hook-form'
-import SignupTitle from './signup-title'
-import { WithFormContext } from './type'
-import { withFormContext } from './with-form-context'
+import {
+  SignupTitle,
+  withFormContext,
+  type WithFormContext,
+} from '@components/signup'
 
 function Password({ onNext, methods, title }: WithFormContext) {
   const { register, control } = methods
