+++ apps/web/src/components/signup/step/sex.tsx
@@ -2,9 +2,11 @@ import { Button } from '@yaksok/ui'
 import { cn } from '@yaksok/utils'
 import { forwardRef } from 'react'
 import { useWatch } from 'react-hook-form'
-import SignupTitle from './signup-title'
-import { WithFormContext } from './type'
-import { withFormContext } from './with-form-context'
+import {
+  SignupTitle,
+  withFormContext,
+  type WithFormContext,
+} from '@components/signup'
 
 interface RadioOptionProps extends React.InputHTMLAttributes<HTMLInputElement> {
   label: string
